reorganise georefhelper into its own folder
