+++ exts/cesium.powertools/cesium/powertools/powertools_window.py
@@ -2,7 +2,8 @@ import logging
 import omni.ui as ui
 from typing import Callable, Optional, List
 from cesium.omniverse.ui import CesiumOmniverseDebugWindow
-from .ui.georef_helper_window import CesiumGeorefHelperWindow
+from .georefhelper.georef_helper_window import CesiumGeorefHelperWindow
+
 
 class PowertoolsAction:
     def __init__(self, title: str, action: Callable):
@@ -35,7 +36,7 @@ class CesiumPowertoolsWindow(ui.Window):
         #  in another window or class, you can absolutely call that instead from here.
         self._actions: List[PowertoolsAction] = [
             PowertoolsAction("Open Cesium Debugging Window", CesiumOmniverseDebugWindow.show_window),
-            PowertoolsAction("Open Cesium Georeference Helper Window", CesiumGeorefHelperWindow.create_window)
+            PowertoolsAction("Open Cesium Georeference Helper Window", CesiumGeorefHelperWindow.create_window),
         ]
 
         self.frame.set_build_fn(self._build_fn)
